refactor: Move calculateDueDate to utils
